requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
hubspot==0.1.14.dev0
hubspot-api-client==12.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pdfrw==0.4
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
